requirements.txt
cryptography==42.0.2
pycryptodome==3.19.1 